requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
allennlp==2.9.0
allennlp-models==2.9.0
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.csv==1.0.7
backports.functools-lru-cache==1.6.4
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
blis==0.7.6
boto3==1.21.22
botocore==1.24.22
brotlipy==0.7.0
cached-path==1.1.0
cached-property==1.5.2
cachetools==5.0.0
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
checklist==0.0.11
cheroot==8.6.0
CherryPy==18.6.1
click==8.0.4
colorama==0.4.4
configparser==5.2.0
conllu==4.4.1
cryptography==36.0.1
cymem==2.0.6
dataclasses==0.8
datasets==2.0.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
fairscale==0.4.5
feedparser==6.0.8
filelock==3.4.2
flit_core==3.7.1
frozenlist==1.3.0
fsspec==2022.2.0
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.5.0
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.1.2
google-resumable-media==2.1.0
googleapis-common-protos==1.56.0
gql==2.0.0
graphql-core==2.3.2
grpcio==1.44.0
h5py==3.6.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
iso-639==0.4.5
jaraco.classes==3.2.1
jaraco.collections==3.5.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.7.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.1.0
jsonnet==0.17.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
langcodes==3.3.0
lmdb==1.3.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==8.12.0
multidict==6.0.2
multiprocess==0.70.12.2
munch==2.5.0
murmurhash==1.0.6
mysqlclient==2.0.3
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
nltk==3.6.5
notebook==6.4.10
numpy==1.22.3
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
overrides==6.1.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.6.1
Pattern==3.6
pdfminer.six==20220319
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==22.0.4
pluggy==1.0.0
portend==2.7.1
preshed==3.0.6
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.27
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-rouge==1.1
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pydantic==1.8.2
Pygments==2.11.2
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
pyu2f==0.1.5
PyYAML==6.0
pyzmq==22.3.0
regex==2022.3.15
repoze.lru==0.7
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.18.0
Routes==2.5.1
rsa==4.8
Rx==1.6.1
s3transfer==0.5.2
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.5.8
setuptools==59.5.0
sgmllib3k==1.0.0
shellingham==1.4.0
shortuuid==1.0.8
simplejson==3.17.6
six==1.16.0
smart-open==5.2.1
smmap==3.0.5
soupsieve==2.3.1
spacy==3.2.2
spacy-legacy==3.0.9
spacy-loggers==1.0.1
srsly==2.4.2
stack-data==0.2.0
subprocess32==3.5.4
tempora==5.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
thinc==8.0.15
threadpoolctl==3.1.0
tokenizers==0.10.3
tomli==2.0.1
torch==1.10.0
torchvision==0.11.0a0+7ecba96
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
transformers==4.15.0
typer==0.4.0
typing_extensions==4.1.1
typing-utils==0.1.0
urllib3==1.26.9
wandb==0.12.1
wasabi==0.9.0
watchdog==0.10.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.6.0
word2number==1.1
xxhash==0.0.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.7.0
